Added a dataset class.
This class abstracts away the handling of the actual data we have.
The best part about it is that we can pass different callbacks
into the dataset class so we can parse our raw data in different ways!
This makes it way more versatile.

Additionally, d3 will probably be implemented here.