Created dummy data and serialized script inputs.
So apparently, when running multiple Python scripts, the context
of the scripts are saved!! Which is a good thing because we
can create modules we can reuse, but also bad because imports
don't work when they repeat the same variable name... so I serialized
all input names to make sure they work regardless.

I also created 345 files worth of dummy data (not included here).