Added a means for exiting app using a button; also added client.ops.js.
I figured that using ctrl+W (which is what I always use to exit apps)
would not be obvious for other users; thus, I added a button to permit
the user to exit the app.

client.ops.js contains miscellaneous requests for services (actions)
that need to be done by the main process but are called by the
renderer.