Upgraded the context api.
When performing updates that aren't meant to trigger rerenders,
the functions that do those updates are passed a copy of the state.
However, because of the way JS works, those objects are passed by value,
EXCEPT for the objects inside them. This is good news because we can simply
encapsulate our state within another object (we call it '_' in our case)
to make sure it's always passed by reference. That allows us to update our
state without using setState().