Updated build config to support pyscripts.
Apparently, files that aren't imported directly using JS are left out
in the package step of the build lifecycle of the app... that meant
that our Python files were not automtically included in the installer
(and thus our imports would break in production).

To fix this, I had to specify to include the files in the package
step of the process, and to properly reference the files, different
file paths were used to import the Python scripts (for dev vs production).